fix: first-round audit fixes — correctness bugs and hot-path performance
Correctness:
- Fix CompiledFieldSlot::tag uint16_t truncation for tags > 65535
- Fix value_length uint16_t truncation for large RawData fields
- Fix AllocateOutboundSeq uint32_t overflow (return error before wrap)
- Fix overlay silently allowing field type override without flag check
- Fix resume_state_ clobbering on second gap during kResendProcessing
- Fix PrecompiledTemplateTable::find() data race on mutable cache
- Merge SaveOutboundView + PersistRecoveryState into single store write
- Replace std::localtime with localtime_r for thread safety

Performance:
- SPSC queue: alignas(64) padding to prevent false sharing
- Eliminate global static mutex + map in encode path (per-session cache)
- Group parsing O(N²) → O(1) tail append with last_entry_index
- Deduplicate find_group() double lookup in decode loop
- Timer wheel: replace O(N) RecomputeEarliestDeadline with min-heap
- MmapStore: incremental index maintenance, eliminate O(N²) full scan
- MmapStore: page-level preallocation to reduce ftruncate frequency

+++ include/fastfix/codec/compiled_decoder.h
@@ -15,7 +15,7 @@ namespace fastfix::codec {
 /// Pre-compiled per-field metadata for a known message type.
 /// Replaces runtime dictionary lookups with direct array indexing.
 struct CompiledFieldSlot {
-    std::uint16_t tag{0};
+    std::uint32_t tag{0};
     message::FieldValueType field_type{message::FieldValueType::kString};
     bool is_group_count{false};
     const profile::GroupDefRecord* group_def{nullptr};
